backend/pyproject.toml
[project]
name = "tablesched-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework",
  "django-tenants",
  "django-allauth",
  "psycopg[binary]",
  "gunicorn",
  "stripe",
  "twilio",
  "django-environ",
  "python-json-logger",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-django", "factory-boy", "responses", "ruff", "mypy"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
